v0.32.4
I'll most likely archive this resource once YAMP is fully released and completed. Alternatively, I may convert it from FiveM to YAMP, as I no longer wish to be involved with the platform. Additionally, I am unsure if the dependencies for this resource will even be maintained, as the creator also plans to transition to YAMP.